app/requirements.txt
faiss_cpu==1.11.0
faiss_cpu==1.12.0
huggingface_hub==0.34.4
joblib==1.5.2
numpy==2.3.3
python-dotenv==1.1.1
scikit_learn==1.7.2
sentence_transformers==5.1.0
streamlit==1.49.1
